pyproject.toml
[project]
name = "demand-forecast"
version = "1.0.0"
description = "Demand forecasting with Transformer neural networks"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "torch>=2.0",
    "pandas>=2.0",
    "numpy>=1.24,<2.0",       # NumPy 2.x has compatibility issues with PyTorch/compiled modules
    "scikit-learn>=1.3",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer[all]>=0.9",
    "tqdm>=4.65",
    "matplotlib>=3.7",
    "torchmetrics>=1.0",
    "torchtnt>=0.2",
    "tsfresh>=0.20",
    "scipy>=1.10",
    "pyyaml>=6.0",
    "joblib>=1.3",
    "dill",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-xdist>=3.3",
    "mypy>=1.5",
    "ruff>=0.1",
    "pre-commit>=3.4",
]

[project.scripts]
demand-forecast = "demand_forecast.cli:app"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["demand_forecast*"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=demand_forecast --cov-report=term-missing"
